Backend/requirements.txt
# ── FastAPI stack ─────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9       # form data parsing

# ── Google / Gmail ─────────────────────────────────────────────────────────
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ── ML / RAG ──────────────────────────────────────────────────────────────
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0              # swap for faiss-gpu if you have CUDA
transformers>=4.41.0
torch>=2.3.0                  # CPU wheel; install torch+cu121 for GPU

# ── HTTP (OpenAI-compat backend) ──────────────────────────────────────────
httpx>=0.27.0

# ── Frontend ──────────────────────────────────────────────────────────────
streamlit>=1.35.0
requests>=2.32.0
